+++ test-results/lambda.c
@@ -14,7 +14,7 @@ void hello_impl_1(struct lambda_ctx_0 *ctx, int num) {
     }
 extern int puts (const char *s);
 extern int printf (const char *format, ...);
-extern void *malloc(int n);
+extern void *malloc(unsigned long n);
 extern void free(void *p);
 extern char *strcpy(char *dest, const char *src);
