+++ test-results/refptr.c
@@ -95,7 +95,7 @@ RefPtr_Node _assign_copy_Node(RefPtr_Node *lhs, RefPtr_Node rhs_val) {
 }
 extern int puts (const char *s);
 extern int printf (const char *format, ...);
-extern void *malloc(int n);
+extern void *malloc(unsigned long n);
 extern void free(void *p);
 extern char *strcpy(char *dest, const char *src);
